rename global_inputs.py to constants.py
