+++ src/pybuildingenergy/constants.py
