+++ src/pybuildingenergy/source/functions.py
@@ -9,7 +9,7 @@ import numpy as np
 import pandas as pd
 from sklearn.metrics import r2_score
 from sklearn.linear_model import LinearRegression
-from pybuildingenergy.global_inputs import main_directory_
+from pybuildingenergy.constants import main_directory_
 import pickle
